+++ test/pebble_test.exs
@@ -6,7 +6,7 @@ defmodule SiteEncrypt.PebbleTest do
   @moduletag :pebble
 
   setup_all do
-    start_supervised!({SiteEncrypt.Phoenix, TestEndpoint})
+    start_supervised!(TestEndpoint)
     :ok
   end
 
@@ -30,19 +30,13 @@ defmodule SiteEncrypt.PebbleTest do
   defmodule TestEndpoint do
     @moduledoc false
 
-    use Phoenix.Endpoint, otp_app: :site_encrypt
-    use SiteEncrypt.Phoenix
-
-    @impl Phoenix.Endpoint
-    def init(_key, config) do
-      {:ok,
-       config
-       |> SiteEncrypt.Phoenix.configure_https(port: 5001)
-       |> Keyword.merge(
-         url: [scheme: "https", host: "localhost", port: 5001],
-         http: [port: 5002]
-       )}
-    end
+    use SiteEncrypt.Phoenix.Endpoint,
+      otp_app: :site_encrypt,
+      endpoint_opts: [
+        http: [port: 5002],
+        https: [port: 5001],
+        url: [scheme: "https", host: "localhost", port: 5001]
+      ]
 
     @impl SiteEncrypt
     def certification do